+++ lib/oneview-sdk/resource/api200/logical_interconnect_group.rb
@@ -40,6 +40,14 @@ module OneviewSDK
         parse_interconnect_map_template if @data['interconnectMapTemplate']['interconnectMapEntryTemplates'] == []
       end
 
+      # Get the logical interconnect group default settings
+      # @param [OneviewSDK::Client] client The client object for the OneView appliance
+      # @return [Hash] The logical interconnect group settings
+      def self.get_default_settings(client)
+        response = client.rest_get(BASE_URI + '/defaultSettings', client.api_version)
+        client.response_handler(response)
+      end
+
       # Adds an interconnect
       # @param [Fixnum] bay Bay number
       # @param [String] type Interconnect type
@@ -64,6 +72,7 @@ module OneviewSDK
       end
 
       # Get the logical interconnect group default settings
+      # @deprecated Use {::get_default_settings} instead. This method will be removed in the next major release.
       # @return [Hash] The logical interconnect group settings
       def get_default_settings
         get_uri = self.class::BASE_URI + '/defaultSettings'