+++ spec/unit/resource/api300/synergy/logical_interconnect_group_spec.rb
@@ -31,7 +31,7 @@ RSpec.describe OneviewSDK::API300::Synergy::LogicalInterconnectGroup do
     end
   end
 
-  describe '#get_default_settings' do
+  describe '::get_default_settings' do
     it 'gets the default settings' do
       expect(@client_300).to receive(:rest_get).with('/rest/logical-interconnect-groups/defaultSettings')
         .and_return(FakeResponse.new('Default' => 'Settings'))
@@ -39,7 +39,7 @@ RSpec.describe OneviewSDK::API300::Synergy::LogicalInterconnectGroup do
     end
   end
 
-  describe '#settings' do
+  describe '#get_settings' do
     it 'gets the current settings' do
       item = described_class.new(@client_300, uri: '/rest/fake')
       expect(@client_300).to receive(:rest_get).with('/rest/fake/settings', 300)